+++ app/src/main/java/com/example/food/utils/Formatters.kt
@@ -3,6 +3,6 @@ package com.example.food.utils
 import java.text.NumberFormat
 import java.util.Locale
 
-fun priceFormatter (price: Int): String {
+fun priceFormatter (price: Float): String {
     return NumberFormat.getCurrencyInstance(Locale.US).format(price)
 }
